Added last set of necessary tests.
